requirements.txt
Pillow
pyfiglet
pywin32
console-progressbar
